perf: reduce heap allocation and improve overall performence (#19)
- **Refactor**
  - Simplified internal handling of job results and result channels for improved efficiency.
  - Updated benchmark worker creation to use default concurrency settings.
  - Enhanced group job synchronization with unified control for better job counting and completion tracking.
  - Improved timing measurement placement in example usage.
  - Adjusted configuration field types for consistency.
- **Documentation**
  - Clarified example usage by explicitly handling errors in queue operations.

### Old bench
```
BenchmarkQueue_Operations/Add-4               	  806361	      1384 ns/op	     384 B/op	       8 allocs/op
BenchmarkQueue_Operations/AddAll-4            	      1194   1003741 ns/op	  243343 B/op	    7008 allocs/op
```

### New bench
```
BenchmarkQueue_Operations/Add-4               	  928090	      1258 ns/op	     320 B/op	       5 allocs/op
BenchmarkQueue_Operations/AddAll-4            	    1416	    833844 ns/op	  235428 B/op	    5005 allocs/op
```

introduce new `Wait` method with group job

+++ config.go
@@ -14,7 +14,7 @@ type configs struct {
 	CleanupCacheInterval     time.Duration
 	JobIdGenerator           func() string
 	IdleWorkerExpiryDuration time.Duration
-	MinIdleWorkerRatio       uint16
+	MinIdleWorkerRatio       uint8
 }
 
 func newConfig() configs {
@@ -58,7 +58,7 @@ func WithIdleWorkerExpiryDuration(duration time.Duration) ConfigFunc {
 // Example: 20 means keep 20% of the concurrency level as idle workers.
 // The actual number of idle workers will be calculated as: concurrency * percentage / 100.
 // The percentage must be between 1 and 100. Values outside this range will be clamped.
-func WithMinIdleWorkerRatio(percentage uint16) ConfigFunc {
+func WithMinIdleWorkerRatio(percentage uint8) ConfigFunc {
 	// Clamp percentage between 1 and 100
 	if percentage == 0 {
 		percentage = 1
